[ENH] - TypeError in `sim/utils.py`
Hey @TomDonoghue , I noticed that this refractory function in sim/utils.py raises a TypeError because this function shares the same name with one of its input arguments refractory period. To fix this, I changed the name of the input argument from refractory to refractory_time.

This may need to be merged before merging its test function (will fail otherwise).